Fix test error on hosts having an old glibc
Since a recent change in Optipng, the CI cannot execute the downloaded version. It's because this version use a too recent glibc library, not available on the CI. To turn around this, I've added a system property disabling the usage of the downloaded version (falling back to system path).

The error is:
[warn] /var/lib/jenkins/jobs/Wisdom-Framework - Build/workspace/core/wisdom-maven-plugin/target/workbench/image/tools/optipng: /lib64/libc.so.6: version `GLIBC_2.14' not found (required by /var/lib/jenkins/jobs/Wisdom-Framework - Build/workspace/core/wisdom-maven-plugin/target/workbench/image/tools/optipng)

Signed-off-by: Clement Escoffier <clement.escoffier@gmail.com>