Fix notion of module in FileWebJarLib and http://localhost:9000/assets rendering.
Add source origin in toString of FileWebJarLib.
FileWebJarLib will only index file associated with a module (source)
when the source jar contain a folder with the same name.

This will ease the merge of webjars with a same source (usefull for webcomponent integration).

Without this commit, webjar with a same source are deployed and merged
but only one webjar is shown in http://localhost:9000/assets